requirements.txt
langchain>=0.2.5
langchain_groq>=0.1.5
streamlit>=1.36.0
watchdog>=4.0.1
python-dotenv>=1.0.1
streamlit-ace>=0.1.1
